requirements.txt

Flask
Flask-CORS
transformers
pdfminer.six
python-docx
torch
requests
python-dotenv



